+++ src/cli/subcommands/load_template_arg.rs
@@ -1,7 +1,7 @@
 use clap::Args;
 
 use crate::cli::init_kind::InitKind;
-use crate::{prelude::*, app_env_name};
+use crate::{app_env_name, prelude::*};
 
 #[derive(Debug, Args, Getters)]
 pub struct LoadTemplateArg {
@@ -10,6 +10,10 @@ pub struct LoadTemplateArg {
     with: InitKind,
     #[arg(long, short, env = app_env_name!("IGNORE_PLACEHOLDERS"))]
     ignore_placeholders: bool,
+    #[arg(long, short,default_value_t = String::from(constants::DEFAULT_LEFT_DELIMITER),  env = app_env_name!("LEFT_DELIMITER"))]
+    left_delimiter: String,
+    #[arg(long, short, default_value_t = String::from(constants::DEFAULT_RIGHT_DELIMITER), env = app_env_name!("RIGHT_DELIMITER"))]
+    right_delimiter: String,
 }
 
 impl LoadTemplateArg {
@@ -18,11 +22,21 @@ impl LoadTemplateArg {
             name,
             with,
             ignore_placeholders: false,
+            left_delimiter: String::from(constants::DEFAULT_LEFT_DELIMITER),
+            right_delimiter: String::from(constants::DEFAULT_RIGHT_DELIMITER),
         }
     }
 
     pub fn activate_ignore_placeholders(mut self) -> Self {
         self.ignore_placeholders = true;
+        self
+    }
+    pub fn new_left_delimiter(mut self, left: String) -> Self {
+        self.left_delimiter = left;
+        self
+    }
+    pub fn new_right_delimiter(mut self, right: String) -> Self {
+        self.right_delimiter = right;
         self
     }
 }