+++ tests/copy_template_into.rs
@@ -3,7 +3,9 @@ use common::prelude::*;
 use crust_boot_you::{
     cli::{InitKind, LoadTemplateArg},
     handle_commands,
-    template_augmentation::{console_fetcher::TestConsoleFetcher, RegexTemplateAugmentor},
+    template_augmentation::{
+        console_fetcher::TestConsoleFetcher, AugementRepository, RegexTemplateAugmentor,
+    },
 };
 use map_macro::hash_map;
 
@@ -202,6 +204,39 @@ fn ignore_templating() {
 
     let mut store: RegexTemplateAugmentor<TestConsoleFetcher> =
         RegexTemplateAugmentor::from_fake(values);
+
+    let output = handle_commands::handle_load_template(
+        setup.path_provider(),
+        setup.os_mani(),
+        &mut store,
+        &arg,
+    )
+    .unwrap();
+
+    setup.assert_with_expected();
+    insta_display_filter_random_tmp!(output);
+}
+#[test]
+#[named]
+fn with_different_placeholder() {
+    let setup = TestSetupBuilder::new(actual_expected!())
+        .suffix_cwd(PathBuf::from("cwd"))
+        .build();
+
+    let deliminter = "%%".to_string();
+    let to_copy_from = ValidTemplateName::new("!!!".to_string()).unwrap();
+    let init_kind = InitKind::Override;
+
+    let arg = LoadTemplateArg::new(to_copy_from, init_kind)
+        .new_left_delimiter(deliminter.clone())
+        .new_right_delimiter(deliminter);
+    let values = hash_map! {"world".to_string() => "P x!!x C".to_string()};
+
+    let console_fetcher = TestConsoleFetcher::new(values);
+    let repo = AugementRepository::new(console_fetcher);
+
+    let mut store: RegexTemplateAugmentor<TestConsoleFetcher> =
+        RegexTemplateAugmentor::from_cli(repo, &arg);
 
     let output = handle_commands::handle_load_template(
         setup.path_provider(),